server/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.0.12
Django==4.1.2
idna==3.4
pycodestyle==2.9.1
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
django-bootstrap3==22.1
django-simpleui~=2022.6.131
django-otp~=1.1.3
qrcode~=7.3.1
djangorestframework~=3.13.1
django-apscheduler~=0.6.2
django-model-utils~=4.3.1
APScheduler~=3.9.1
eth_account~=0.5.9
pyjwt~=2.6.0
pymysql~=1.0.2
mysqlclient~=2.1.1
django-cors-headers~=3.13.0
aws_requests_auth~=0.4.3
google-cloud-logging==3.2.5
google-cloud-bigquery==3.4.0
google-cloud-storage==2.6.0
google-cloud-pubsub==2.13.11
django-import-export~=3.0.1
py-ecc==5.2.0
ssz==0.2.4
pylibmc~=1.6.3
django-prometheus==2.2.0
prometheus-client==0.15.0
web3==5.31.3
